requirements.txt
schema
pyyaml
types-PyYAML
mypy
selenium